# assetmgmt/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Badminton Analysis API Documentation",
        "version": "1.0.0",
        "description": "A Node.js API documentation for Badminton Analysis project",
        "contact": {
            "name": "Mithun Kumar",
            "email": "mithun.k@borqs.com"
        },
        "license": {
            "name": "Architectural Design",
            "url": "/assetmgmt/architecture"
        },
        "termsOfService": "/assetmgmt/architecture/terms-of-service"
    },
    "servers": [
        {
            "url": "http://api.visist.ai/assetmgmt"
        },
        {
            "url": "http://stg.borqs.io/assetmgmt"
        },
        {
            "url": "http://localhost:9002/assetmgmt"
        }
    ],
...
    "components": {
        "schemas": {
...
            "updateGamePost": {
                "type": "object",
                "properties": {
                    "strokeTechnique": {
                        "type": "object",
                        "properties": {
                            "forehand": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "description": "enter the array of forehand stroke techniques"
                                }
                            },
                            "backhand": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "description": "enter the array of backhand stroke techniques"
                                }
                            }
                        }
                    },
                    "footwork": {
                        "type": "object",
                        "properties": {
                            "attack": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "description": "enter the array of attack footwork techniques"
                                }
                            },
                            "defence": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "description": "enter the array of defence footwork techniques"
                                }
                            }
                        }
                    }
                }
            },
            "listPosts": {
                "type": "object",
                "properties": {
                    "condition": {
                        "type": "object",
                        "properties": {
                            "sharedWith": {
                                "type": "string",
                                "description": "enter the ID of the coach with whom the data is shared"
                            }
                        }
                    }
                }
            },
            "listPostsReviewer": {
                "type": "object",
                "properties": {
                    "condition": {
                        "type": "object",
                        "properties": {
                            "state": {
                                "type": "string",
                                "description": "enter the state of the review"
                            },
                            "poster": {
                                "type": "string",
                                "description": "enter the ID of the player who posted the data"
                            }
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/post/?type=game": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "parameters": [
                    {
                        "name": "filename",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "streamsize",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "summary": "create a game post",
                "description": "Create a game post by a user",
                "operationId": "createGamePost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "video/*": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "one post created successfully."
                    },
                    "400": {
                        "description": "authentication error"
                    },
                    "401": {
                        "description": "unauthorized! invalid token. error in verifying token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/?postId={postId}": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "enter the post id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filename",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "streamsize",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "summary": "update a game post",
                "description": "Update a game post by a user",
                "operationId": "updateGamePost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updateGamePost"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "one post updated successfully"
                    },
                    "400": {
                        "description": "authentication error"
                    },
                    "401": {
                        "description": "unauthorized! invalid token. error in verifying token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/post/?postId={postId}&type=game": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "enter the post id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "summary": "fetch a post",
                "description": "Get the content of the post",
                "operationId": "fetchGamePost",
                "responses": {
                    "200": {
                        "description": "file recieved successfully"
                    },
                    "400": {
                        "description": "authentication error"
                    },
                    "401": {
                        "description": "unauthorized! invalid token. error while retrieving one user by id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/?postId={postId}&type=game&file={tnFilePath}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "enter the post id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tnFilePath",
                        "in": "path",
                        "description": "enter the thumbnail path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "summary": "fetch gamepost thumbnail",
                "description": "Get a thumbnail relevant to the post",
                "operationId": "getThumbnail",
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "400": {
                        "description": "authentication error"
                    },
                    "401": {
                        "description": "unauthorized! invalid token. error in verifying token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/?postId={postId}&type=review": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "enter the post id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "summary": "fetch a reviewed post",
                "description": "get a reviewed post contents",
                "operationId": "fetchReviewedPost",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "authentication error"
                    },
                    "401": {
                        "description": "unauthorized! invalid token. error in verifying token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/searchCount": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "coach - get post count shared with me",
                "description": "get the post counts shared with a coach",
                "operationId": "postsCount",
                "responses": {
                    "200": {
                        "description": "posts count retrieved successfully"
                    },
                    "400": {
                        "description": "authentication error"
                    },
                    "401": {
                        "description": "unauthorized! invalid token. error in verifying token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/search": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "coach - list all posts shared with me",
                "description": "get all the posts shared with a coach",
                "operationId": "listAllPosts",
                "requestBody": {
                    "required": "true",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/listPosts"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "posts retrieved successfully"
                    },
                    "400": {
                        "description": "authentication error"
                    },
                    "401": {
                        "description": "unauthorized! invalid token. error in verifying token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/search/": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "reviewer - list all posts related with a poster",
                "description": "get all the posts listed under a user with userId",
                "operationId": "listAllPostsForReviewer",
                "requestBody": {
                    "required": "true",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/listPostsReviewer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "posts retrieved successfully"
                    },
                    "400": {
                        "description": "authentication error"
                    },
                    "401": {
                        "description": "unauthorized! invalid token. error in verifying token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/searchCount/": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "reviewer - list all posts count related with a poster",
                "description": "get all the posts count listed under a user with userId",
                "operationId": "listAllPostsForReviewer",
                "requestBody": {
                    "required": "true",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/listPostsReviewer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "posts count retrieved successfully"
                    },
                    "400": {
                        "description": "authentication error"
                    },
                    "401": {
                        "description": "unauthorized! invalid token. error in verifying token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/?type=review&gamePost={postId}": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filename",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "streamsize",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "reviewindex",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "summary": "reviewer - create a review post",
                "description": "reviewer - create a review post",
                "operationId": "updateProfilePic",
                "requestBody": {
                    "required": true,
                    "content": {
                        "video/*": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "one post updated successfully"
                    },
                    "400": {
                        "description": "authentication error"
                    },
                    "401": {
                        "description": "unauthorized! invalid token. error in verifying token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/?postId={postId}/": {
            "delete": {
                "tags": [
                    "Posts"
                ],
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "summary": "delete a post by id",
                "description": "delete a post by id",
                "operationId": "deletePost",
                "responses": {
                    "200": {
                        "description": "one post deleted successfully by id"
                    },
                    "400": {
                        "description": "authentication error"
                    },
                    "401": {
                        "description": "unauthorized! invalid token. error in verifying token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}